Refactor VerticalReaderView for improved chapter navigation and state management
- Updated VerticalReaderView to use chapter IDs instead of sequences for better clarity.
- Replaced ChapterStore with ValueNotifiers for chapter loading states.
- Simplified scroll handling and position reporting logic.
- Removed unnecessary item rebuilding logic and resolved items structure.
- Enhanced error handling in Rust code with new error constants.
- Modified ParagraphId to use u64 for indexing.
- Implemented unique ID assignment in TransformOutput to prevent duplicates.
- Added tests for paragraph ID uniqueness in TransformOutput.

+++ native/langhuan/src/model.rs
@@ -91,7 +91,7 @@ impl ChapterInfo {
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
 #[serde(untagged)]
 pub enum ParagraphId {
-    Index(usize),
+    Index(u64),
     Id(String),
 }
 
@@ -183,7 +183,7 @@ impl RawParagraph {
     pub(crate) fn into_paragraph(self, index: usize) -> Paragraph {
         let make_id = |opt: Option<String>| match opt {
             Some(s) => ParagraphId::Id(s),
-            None => ParagraphId::Index(index),
+            None => ParagraphId::Index(index as u64),
         };
         match self {
             RawParagraph::Title { id, text } => Paragraph::Title {
